Update June 11, 2019
Added Models for checklist and courses,
Updated courses page,
Added table plugin for TinyMCE,
Fixed display for date and time for articles,
Fixed time to show correctly as it showed UTC+0 before.

Checklists are stored as MEDIUMTEXT where it contains html source code created by the TinyMCE editor. As of now, the editor does not have integration with Bootstrap components.